+++ doc/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2016/1/11 1.2.1 ##
+
+Changed
+ - Cambiate alcune variabili da short a unsigned char
+
+Fixed
+ - Vengono controllati i caratteri della stringa, prima di passarla alle funzioni
+   per verificare se è una password valida
+
 ## 2016/1/11 1.1.0 ##
 
 Changed